fix: prefer web font over local font
Since preload is added, it's a waste not to use it.

This solves the problem that the local installed font could be broken,
e.g. no bold style, no halt feature, etc.

BTW, now kai fallbacks to serif instead of --default-serif-font, as it's
not much better to fallback to --default-serif-font instead of serif,
and KaiTi is widely available.